backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
yfinance
google-genai
feedparser
httpx
pydantic
pydantic-settings
python-dotenv
aiosqlite
